Coerce the LLFF near/far bounds to plain floats
In the no_ndc branch these came straight from np.ndarray.min/max on the bds
array, so they were numpy scalars rather than Python floats. Every other branch
assigns literals (0., 1., or the blender/LINEMOD constants), making this the one
path where near and far had a different type.

They flow into the bds_dict that render_kwargs carries, and from there into
render() and render_rays(), where they are broadcast against tensors to build
z_vals. A numpy float64 scalar works there, but under NEP 50 it can promote the
float32 ray math to float64 -- the same trap 025a476 addressed for K in
get_rays_np. float() keeps the arithmetic at the dtype the rest of the pipeline
uses.

Only the llff dataset with --no_ndc reaches this code, and no LLFF data is
present here, so this is verified by construction rather than by a run: the two
expressions are unchanged apart from the wrapping, and float() on a 0-d numpy
scalar preserves its value.

+++ run_nerf.py
@@ -583,8 +583,8 @@ def train():
 
         print('DEFINING BOUNDS')
         if args.no_ndc:
-            near = np.ndarray.min(bds) * 0.9
-            far = np.ndarray.max(bds) * 1.
+            near = float(np.ndarray.min(bds) * 0.9)
+            far = float(np.ndarray.max(bds) * 1.)
 
         else:
             near = 0.
